+++ packages/react-dev-utils/package.json
@@ -24,12 +24,13 @@
     "ModuleScopePlugin.js",
     "openBrowser.js",
     "openChrome.applescript",
-    "prepareProxy.js",
     "printHostingInstructions.js",
     "WatchMissingNodeModulesPlugin.js",
+    "WebpackDevServerUtils.js",
     "webpackHotDevClient.js"
   ],
   "dependencies": {
+    "@timer/detect-port": "1.1.3",
     "address": "1.0.1",
     "anser": "1.3.0",
     "babel-code-frame": "6.22.0",
@@ -39,6 +40,7 @@
     "filesize": "3.3.0",
     "gzip-size": "3.0.0",
     "html-entities": "1.2.1",
+    "inquirer": "3.0.6",
     "opn": "5.0.0",
     "recursive-readdir": "2.2.1",
     "shell-quote": "1.6.1",
