+++ packages/react-scripts/package.json
@@ -21,8 +21,6 @@
     "react-scripts": "./bin/react-scripts.js"
   },
   "dependencies": {
-    "@timer/detect-port": "1.1.3",
-    "address": "1.0.1",
     "autoprefixer": "7.1.0",
     "babel-core": "6.24.1",
     "babel-eslint": "7.2.3",
